requirements.txt
flask~=2.2.2
python-dotenv~=0.21.1
flask-api
flask-pymongo
haralyzer~=2.2.0
ndjson~=0.3.1
requests~=2.28.2
pycountry~=22.3.5
langdetect~=1.0.9
pytesseract~=0.3.10
pillow~=9.4.0
beautifulsoup4~=4.11.2